// apps/packages/app/src/websocket/interface.ts
export enum WebsocketSharedWorkerEnum {
	WISE_NAME = 'olares-wise-socket-worker',
	FILES_NAME = 'olares-files-socket-worker',
	MARKET_NAME = 'olares-market-socket-worker',
	SETTINGS_NAME = 'olares-settings-socket-worker',
	VAULT_NAME = 'olares-vault-socket-worker',
	DESKTOP_NAME = 'olares-desktop-socket-worker',
	DASHBOARD_NAME = 'olares-dashboard-socket-worker',
	STUDIO_NAME = 'olares-studio-socket-worker'
}

export enum WebsocketApplicationEnum {
	LAREPASS = 'larepass',
	LarePass_WISE = 'larepass_wise'
	// VAULT = 'vault',
	// DASHBOARD = 'dashboard',
	// MARKET = 'market',
	// DESKTOP = 'desktop'
	// SETTINGS = 'settings'
}